Add subscription model

+++ src/utils/cloudinary.js
@@ -1,3 +1,6 @@
+import dotenv from "dotenv";
+dotenv.config();
+
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs";
 
@@ -10,18 +13,30 @@ cloudinary.config({
 
 // Upload an image
 export const uploadOnCloudinary = async (localFilePath) => {
+  if (!localFilePath) return null;
+
   try {
-    if (!localFilePath) return null;
-    //upload the file on cloudinary
+    // Upload the file on Cloudinary
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
-    // file has been uploaded successfull
-    //console.log("file is uploaded on cloudinary ", response.url);
+
+    // File has been uploaded successfully
+    // console.log("file is uploaded on cloudinary ", response.url);
+
+    // Remove the locally saved temporary file
     fs.unlinkSync(localFilePath);
+
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
-    return null;
+    // Log the error
+    console.error("Error uploading to Cloudinary:", error);
+
+    // Remove the locally saved temporary file if it exists
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    }
+
+    return { error: error.message || "An error occurred during the upload" };
   }
 };
